meta crate sources reorg
