+++ framework/meta/src/cmd/contract/generate_snippets/mod.rs
