+++ framework/meta/src/cmd/contract/generate_snippets/snippet_crate_gen.rs
