+++ framework/meta/src/cmd/contract/generate_snippets/snippet_gen_common.rs
