+++ framework/meta/src/cmd/contract/mod.rs
@@ -1,0 +1,8 @@
+mod cli_contract;
+mod generate_snippets;
+mod meta_abi;
+mod meta_config;
+mod output_contract;
+mod validate_abi;
+
+pub use cli_contract::{cli_main, multi_contract_config};
