+++ framework/meta/src/cmd/mod.rs
@@ -1,0 +1,2 @@
+pub mod contract;
+pub mod standalone;
