+++ framework/meta/src/cmd/standalone/mod.rs
@@ -1,0 +1,6 @@
+mod all;
+mod cli_standalone;
+mod info;
+mod upgrade;
+
+pub use cli_standalone::cli_main_standalone;
